Moved driver_operation_bloc into the feature folder
