+++ lib/features/driver/presentation/bloc/driver_operations_bloc/driver_operations_repository/dtos/driver_connect_dto.freezed.dart
