+++ lib/features/driver/presentation/bloc/driver_operations_bloc/driver_operations_repository/dtos/driver_trip_request.dto.dart
@@ -1,5 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:malawi_ride_share_app/app_blocs/driver_operations_bloc/driver_operations_repository/dtos/location_dto.dart';
+import 'package:malawi_ride_share_app/features/driver/presentation/bloc/driver_operations_bloc/driver_operations_repository/dtos/location_dto.dart';
 
 part 'driver_trip_request.dto.freezed.dart';
 part 'driver_trip_request.dto.g.dart';
