+++ lib/features/driver/presentation/bloc/driver_operations_bloc/driver_operations_repository/dtos/driver_trip_request.dto.freezed.dart
